refactor(frontend): move owner check to parent component
